changed file structure and added some content in ContainerSQL nd BaggageSQL
