+++ components/persistence/src/main/java/engine/LogEngine.java
@@ -1,3 +1,7 @@
+package engine;
+
+import main.Configuration;
+
 import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
