+++ components/persistence/src/main/java/main/IDatabase.java
@@ -1,3 +1,5 @@
+package main;
+
 import base.*;
 
 import java.util.ArrayList;
